backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
Pillow==10.1.0
dj-database-url==2.1.0
